pyproject.toml
[project]
name = "databricks-app"
version = "0.1.0"
description = "A Databricks App built with vibe coding"
readme = "README.md"
requires-python = ">=3.10"

# Core dependencies - installed in Databricks App runtime
# Reference: https://github.com/databricks-solutions/claude-databricks-app-template
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",  # For file uploads
    
    # Databricks Integration
    "databricks-sdk>=0.20.0",
    "mlflow[databricks]>=2.10.0",  # ML tracking, model serving, AI agents
    
    # Data Processing
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",  # For settings management
    
    # HTTP & Async
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",  # Structured logging
]

# Development-only dependencies (not deployed to Databricks)
[project.optional-dependencies]
dev = [
    # Local development with Databricks
    "databricks-connect>=14.0.0",
    
    # Code Quality
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    
    # Type Checking
    "mypy>=1.0.0",
    
    # API Client Generation
    "openapi-python-client>=0.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Ruff Configuration (Python linter/formatter)
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

# MyPy Configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
